core: move files to core
